feat: implement multi-tenant core services, repository infrastructure, and database schema migration for store isolation

+++ backend/src/main/java/com/sciencebot/pos/settings/internal/repositories/SettingsRepository.java
@@ -3,5 +3,5 @@ package com.sciencebot.pos.settings.internal.repositories;
 import com.sciencebot.pos.settings.internal.entities.Setting;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SettingsRepository extends JpaRepository<Setting, Integer> {
+public interface SettingsRepository extends JpaRepository<Setting, Long> {
 }
